% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LC2012.R
\docType{data}
\name{LC2012}
\alias{LC2012}
\title{2012 Liberal-Conservative Scales}
\format{
An object of class \code{data.frame} with 5914 rows and 6 columns.
}
\source{
American National Election Studies. \href{http://www.electionstudies.org/}{www.electionstudies.org/}
}
\usage{
data(LC2012)
}
\description{
Liberal-Conservative 7-point scales from the 2012 National Election Study. Includes
(in order) original case id, self-placement, and rankings of Obama, Romney, Democratic Party,
Republican party. Missing values are coded as NA.
}
\examples{
data(LC2012)
head(LC2012)
}
\keyword{datasets}
