% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_over_self.R
\name{plot_over_self}
\alias{plot_over_self}
\title{Plot individual parameter estimates over self-placements}
\usage{
plot_over_self(
  objects,
  data,
  par = "chi",
  estimate = "median",
  names = NULL,
  parlabel = NULL,
  fill = "#2166AC",
  color = "#053061",
  width = 0.7,
  alpha = 0.5,
  outlier.size = 0.3,
  median_color = "black",
  median_linewidth = 0.7
)
}
\arguments{
\item{data}{The list of data that was used to produce the \code{stanfit} object(s).}

\item{par}{Character: Name of the parameter to be plotted. One of the following: \code{"alpha"}, \code{"beta"}, \code{"abs_beta"}, \code{"lambda"}, \code{"chi"}, and \code{"eta"}. Defaults to \code{"chi"}.}

\item{estimate}{Character: Specifying which type of posterior point estimate to use. One of \code{"median"} and \code{"mean"}. Defaults to \code{"median"}.}

\item{names}{An optional character vector of model names of same length as the supplied list of models.}

\item{parlabel}{An optional character containing an alternative label for the parameter (will be parsed if passed as an expression).}

\item{fill}{Fill color of boxes. Passed on to \code{geom_boxplot}.}

\item{color}{Color of outer lines. Passed on to \code{geom_boxplot}.}

\item{width}{Width of boxes. Passed on to \code{geom_boxplot}.}

\item{alpha}{Number in [0,1]: Inverse level of transparency for fill color. Passed on to \code{geom_boxplot}.}

\item{outlier.size}{Size of dots representing outliers. Passed on to \code{geom_boxplot}.}

\item{median_color}{Color of solid line representing the median.}

\item{median_linewidth}{Thickness of solid line representing the median.}

\item{object}{An object of class \code{stanfit} produced by \code{hbam}, or a \code{list} of such objects, which will produce a faceted plot.}
}
\value{
A \code{ggplot} object.
}
\description{
Create a boxplot of individual parameter estimates from an HBAM model over self-placements
}
\examples{
# Loading and recoding data
data(LC1980)
dat <- LC1980
dat[dat == 0 | dat == 8 | dat == 9] <- NA
self <- dat[, 1]
stimuli <- dat[, -1]

# Preparing data and fitting model:
dat <- prep_data(self, stimuli)
fit_hbam <- hbam(data = dat, prep_data = FALSE)

# Plot estimated respondent positions over self-placements:
plot_over_self(fit_hbam, dat, "chi")
}
