% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_stimuli.R
\name{plot_stimuli}
\alias{plot_stimuli}
\title{Plot estimated stimulus positions}
\usage{
plot_stimuli(object, rev_color = FALSE, alpha = 0.55)
}
\arguments{
\item{object}{An instance of class \code{stanfit} produced by \code{hbam}.}

\item{rev_color}{Logical: Display low positions as red and high positions as blue.}

\item{alpha}{Number in [0,1]: Inverse level of transparency for fill color.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot marginal posterior distributions of stimulus positions from an HBAM model
}
