% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_code.R
\name{show_code}
\alias{show_code}
\title{Show the code for an HBAM or FBAM model}
\usage{
show_code(model = NULL)
}
\arguments{
\item{model}{Character: Name of the model. See the documentation for the \code{hbam()} function for a list of the available models.}
}
\value{
The function prints Stan code.
}
\description{
Show the Stan code for one of the models in the package.
}
\examples{
show_code("HBAM")
}
