% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LC1980.R
\docType{data}
\name{LC1980}
\alias{LC1980}
\title{1980 Liberal-Conservative Scales}
\format{
An object of class \code{matrix} (inherits from \code{array}) with 888 rows and 7 columns.
}
\source{
American National Election Studies: \href{https://www.electionstudies.org/}{www.electionstudies.org/}.

This dataset was originally part of the \code{basicspace} package under the same name ("LC1980").
}
\usage{
data(LC1980)
}
\description{
Liberal-Conservative 7-point scales from the 1980 National Election Study.
Includes (in order) self-placement, and rankings of Carter, Reagan, Kennedy,
Anderson, Republican party, Democratic Party. Stored as a matrix of integers.
The numbers 0, 8, and 9 are considered to be missing values.
}
\examples{
data(LC1980)
dat <- LC1980
dat[dat == 0 | dat == 8 | dat == 9] <- NA
head(dat)
}
\keyword{datasets}
