% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_by_group.R
\name{plot_by_group}
\alias{plot_by_group}
\title{Plot posterior densities of parameter averages by group}
\usage{
plot_by_group(
  object,
  data,
  group_id,
  par = "abs_beta",
  fill = "#2166AC",
  color = "#053061",
  alpha = 0.5,
  ncol = max(1, round(length(unique(group_id))/10)),
  ascending_means = TRUE
)
}
\arguments{
\item{object}{An instance of class \code{stanfit} produced by \code{hbam()}.}

\item{data}{The list of data that was used to produce the object.}

\item{group_id}{A vector that will be used to split the respondents into groups before plotting. The vector must either be as long as the number of rows in the original dataset, or as long as the number of respondents included in the analysis. The \code{group_id} supplied here can be the same as the \code{group_id} argument used when fitting a "MULTI"-type model, but it does not have to be.}

\item{par}{Character: Name of the parameter to be plotted. One of the following: \code{"alpha"}, \code{"beta"}, \code{"abs_beta"}, \code{"lambda"}, or \code{"chi"}. Defaults to \code{"abs_beta"}, which means the absolute values of the draws for beta will be used. Further individual-level parameters like \code{"eta"} can be specified if these have been passed to \code{hbam()} via the argument \code{extra_pars} when fitting the model. (Note that homoskedastic models have no \code{"eta"} parameters and "NF"-type models have no \code{"lambda"} or \code{"kappa"} parameters.)}

\item{fill}{Fill color. Passed on to \code{ggplot2::geom_density()}.}

\item{color}{Color of outer lines. Passed on to \code{ggplot2::geom_density()}.}

\item{alpha}{Number in [0,1]: Inverse level of transparency.}

\item{ncol}{Number of columns. The default uses a formula to have approximately ten subplots per column.}

\item{ascending_means}{Logical: Should the groups be placed in ascending order based on their posterior means (\code{TRUE}) or should they be ordered based on their names (\code{FALSE})? Defaults to \code{TRUE}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot posterior densities of group summaries of individual parameters. The respondents can be grouped by any categorical variable and the function works whether the fitted model is of "MULTI"-type or not.
}
